Now the document field is required in all issues.
This is the first step in the new workflow. Verification of the uploaded
files is made on the creation of the issue on the database, as long as
the upload was sucessfull.
Field verify is made in: BUG:report -> bug_report_page.php -> include ->
bug_report.php -> 200: create() from bug_api.php -> 464:validate() from
bug_api.php - > $this (files)

Also the layout was updated to show the requirement.
And a new option menu was added for the document verification fase.